moved _check_number() function from _fcm_class.py to models.py
