+++ fcm/_fcm_class.py
@@ -1,7 +1,4 @@
-# main class interface
-__all__ = ["StructuralGroup", "FCMmeteoroid", "PancakeMeteoroid", "FragmentationMeteoroid",
-           "FCMparameters", "simulate_impact", "default_cratering_params",
-           "default_cloud_dispersion_coeff"]
+"""Fragment-cloud classes and functions for interacting with the C++ code"""
 
 import math
 from collections.abc import Iterable
@@ -10,39 +7,7 @@ import pandas as pd
 
 from . import _cpp_wrapper as cpp
 from .models import *
-
-
-###################################################
-def _check_number(x, name, check_float=True, lower_bound=None, lower_incl=False,
-                  upper_bound=None, upper_incl=False, bound_format="{:.0f}"):
-    
-    assert isinstance(name, str), "Bug: variable name must be a string"
-    if check_float:
-        if not isinstance(x, (int, float)):
-            raise TypeError("{} must be a float".format(name))
-    else:
-        if not isinstance(x, int):
-            raise TypeError("{} must be an int".format(name))
-    
-    if lower_bound is not None:
-        assert isinstance(lower_bound, (int, float)), "Bug: lower bound must have type int or float"
-        if lower_incl:
-            if x < lower_bound:
-                raise ValueError(("{} must be >= " + bound_format).format(name, lower_bound))
-        elif x <= lower_bound:
-            raise ValueError(("{} must be > " + bound_format).format(name, lower_bound))
-    
-    if upper_bound is not None:
-        assert isinstance(upper_bound, (int, float)), "Bug: upper bound must have type int or float"
-        if upper_incl:
-            if x > upper_bound:
-                raise ValueError(("{} must be <= " + bound_format).format(name, upper_bound))
-        elif x >= upper_bound:
-            raise ValueError(("{} must be < " + bound_format).format(name, upper_bound))
-    
-    if check_float:
-        return float(x)
-    return int(x)
+from .models import _check_number
 
 
 ###################################################
